fix hyperkube to run from goland.
