+++ cmd/service-catalog/main.go
@@ -22,6 +22,7 @@ package main
 import (
 	"os"
 
+	"github.com/kubernetes-incubator/service-catalog/cmd/service-catalog/server"
 	"github.com/kubernetes-incubator/service-catalog/pkg/hyperkube"
 )
 
@@ -31,8 +32,8 @@ func main() {
 		Long: "This is an all-in-one binary that can run any of the various Kubernetes service-catalog servers.",
 	}
 
-	hk.AddServer(NewAPIServer())
-	hk.AddServer(NewControllerManager())
+	hk.AddServer(server.NewAPIServer())
+	hk.AddServer(server.NewControllerManager())
 
 	hk.RunToExit(os.Args)
 }
